requirements.txt
line-bot-sdk==2.3.0
flask==2.1.2
requests==2.28.0
gunicorn
Werkzeug==2.0.3
redis
openai
pymysql
numpy
schedule


